[REF] Replace Promise.all with async/await for role creation
This commit refactors the code to replace the use of "Promise.all" with "async/await" for creating roles. By defining an "async" function "createRoles", we can use "await" inside it to wait for the "Role.create" promises to resolve. This simplifies the code and makes it easier to read and maintain.

The refactored code is still fully functional and can replace the old code completely.